Extract common classes to common package.
