+++ lens4j-processor/src/test/resources/cases/decapitalize-factory-name/AccountWithDeCapitalizeSpecificFactoryName.java
@@ -1,5 +1,6 @@
 package util.examples;
 
+import common.Currency;
 import dev.khbd.lens4j.core.annotations.GenLenses;
 import dev.khbd.lens4j.core.annotations.Lens;
 import dev.khbd.lens4j.core.annotations.LensType;
