+++ lens4j-processor/src/test/resources/cases/milty-property/Payment.java
@@ -1,5 +1,8 @@
 package util.examples;
 
+import common.Currency;
+import common.Receiver;
+import common.Payer;
 import dev.khbd.lens4j.core.annotations.GenLenses;
 import dev.khbd.lens4j.core.annotations.Lens;
 import dev.khbd.lens4j.core.annotations.LensType;
