+++ lens4j-processor/src/test/resources/common/Bank.java
@@ -1,4 +1,4 @@
-package util.examples;
+package common;
 
 public class Bank {
     private Integer id;
