+++ lens4j-processor/src/test/resources/common/Currency.java
@@ -1,4 +1,4 @@
-package util.examples;
+package common;
 
 public class Currency {
     private Integer id;
